moved initial_state.py up one directory
