+++ compass/ocean/tests/overflow/nonhydro/__init__.py
@@ -1,5 +1,5 @@
+from compass.ocean.tests.overflow.initial_state import InitialState
 from compass.ocean.tests.overflow.nonhydro.forward import Forward
-from compass.ocean.tests.overflow.nonhydro.initial_state import InitialState
 from compass.testcase import TestCase
