+++ compass/ocean/tests/overflow/nonhydro/nonhydro.cfg
@@ -27,36 +27,3 @@ coord_type = z-star
 
 # Whether to use "partial" or "full", or "None" to not alter the topography
 partial_cell_type = None
-
-# Config options for the solitary wave testcase
-[hydro_vs_nonhydro]
-
-# Cold water range
-cold_water_range = 990.0
-
-# Lateral position of the shelf-break (m)
-xs = 4700.0
-
-# Length-scale of the slope (m)
-Ls = 533.3333
-
-# Shelf depth (m)
-shelf_depth = 40.0
-
-# Beta in eos
-eos_linear_beta = 0.8
-
-# Reference salinity (PSU)
-eos_linear_Sref = 35.0
-
-# Lower temperature (deg C)
-lower_temperature = 10.0
-
-# Higher temperature (deg C)
-higher_temperature = 20.0
-
-# Config options for the solitary wave testcase
-[visualize]
-
-# Time index at which to plot the variables
-plot_time = 6
