+++ compass/ocean/tests/overflow/overflow.cfg
@@ -9,3 +9,60 @@ length = 200
 
 # Viscosity values to test for rpe test case
 viscosities = 1, 5, 10, 100, 1000
+
+# Config options for the horizontal grid
+[horizontal_grid]
+
+# Number of cells in the x-direction
+nx = 320
+
+# Number of cells in the y-direction
+ny = 4
+
+# Distance from two cell centers
+dc = 20.0
+
+# Options related to the vertical grid
+[vertical_grid]
+
+# Depth of the bottom of the ocean (m)
+bottom_depth = 200.0
+
+# Number of vertical levels
+vert_levels = 60
+
+# The type of vertical grid
+grid_type = uniform
+
+# The type of vertical coordinate (e.g. z-level, z-star)
+coord_type = z-star
+
+# Whether to use "partial" or "full", or "None" to not alter the topography
+partial_cell_type = None
+
+# Config options for the nonhydro test cases
+[hydro_vs_nonhydro]
+
+# Cold water range
+cold_water_range = 990.0
+
+# Lateral position of the shelf-break (m)
+xs = 4700.0
+
+# Length-scale of the slope (m)
+Ls = 533.3333
+
+# Shelf depth (m)
+shelf_depth = 40.0
+
+# Beta in eos
+eos_linear_beta = 0.8
+
+# Reference salinity (PSU)
+eos_linear_Sref = 35.0
+
+# Lower temperature (deg C)
+lower_temperature = 10.0
+
+# Higher temperature (deg C)
+higher_temperature = 20.0
